Allow playing tracks longer than 6 h

+++ src/waveform/renderers/glwaveformrenderersimplesignal.cpp
@@ -47,7 +47,7 @@ void GLWaveformRendererSimpleSignal::draw(QPainter* painter, QPaintEvent* /*even
         return;
     }
 
-    const int trackSamples = m_waveformRenderer->getTrackSamples();
+    const double trackSamples = m_waveformRenderer->getTrackSamples();
     if (trackSamples <= 0) {
         return;
     }
