+++ src/waveform/renderers/waveformrenderbeat.cpp
@@ -42,7 +42,7 @@ void WaveformRenderBeat::draw(QPainter* painter, QPaintEvent* /*event*/) {
     }
     m_beatColor.setAlphaF(alpha/100.0);
 
-    const int trackSamples = m_waveformRenderer->getTrackSamples();
+    const double trackSamples = m_waveformRenderer->getTrackSamples();
     if (trackSamples <= 0) {
         return;
     }
